aws_monitoring/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.16
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
twilio==9.0.2
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
